+++ src/main.ts
@@ -4,7 +4,7 @@ import { originalAddEventListener, originalFetch, originalNodeAppendChild, origi
 import { i18n } from "./i18n";
 import { DownloadType, isPageType, MessageType, PageType, ToastType, VersionState } from "./type";
 import { config, Config } from "./config";
-import { Dictionary, SyncDictionary, Version, VideoInfo } from "./class";
+import { Dictionary, SyncDictionary, Version, VideoInfo, TaskController } from "./class";
 import { db } from "./db";
 import "./date";
 import { delay, findElement, renderNode, unlimitedFetch } from "./extension";
@@ -751,54 +751,7 @@ async function addDownloadTask() {
     unsafeWindow.document.body.appendChild(body)
 }
 
-class TaskController {
-    private lastTaskStartTime: number = 0;
-    private counter: number = 0;
-    private waiting: Array<() => Promise<void>> = [];
-    private isProcessing: boolean = false;
 
-    constructor(private maxConcurrent: number, private minInterval: number) { }
-
-    addTask(task: () => Promise<void>) {
-        this.waiting.push(task);
-        GM_getValue('isDebug') && console.debug(`TaskController: Task added to queue. Queue length: ${this.waiting.length}`);
-        this.processQueue();
-    }
-
-    private async processQueue() {
-        if (this.isProcessing) return;
-        this.isProcessing = true;
-
-        while (this.waiting.length > 0) {
-            if (this.counter >= this.maxConcurrent) {
-                GM_getValue('isDebug') && console.debug(`TaskController: Max concurrent tasks (${this.maxConcurrent}) reached. Pausing queue processing.`);
-                this.isProcessing = false;
-                return;
-            }
-
-            const now = Date.now();
-            const timeSinceLastTask = now - this.lastTaskStartTime;
-            if (timeSinceLastTask < this.minInterval) {
-                GM_getValue('isDebug') && console.debug(`TaskController: Waiting ${this.minInterval - timeSinceLastTask}ms to maintain minimum interval`);
-                await delay(this.minInterval - timeSinceLastTask);
-            }
-
-            const task = this.waiting.shift();
-            if (task) {
-                this.counter++;
-                this.lastTaskStartTime = Date.now();
-                GM_getValue('isDebug') && console.debug(`TaskController: Starting task. Active tasks: ${this.counter}, Remaining in queue: ${this.waiting.length}`);
-
-                task().finally(() => {
-                    this.counter--;
-                    GM_getValue('isDebug') && console.debug(`TaskController: Task completed. Active tasks: ${this.counter}, Remaining in queue: ${this.waiting.length}`);
-                    this.processQueue();
-                });
-            }
-        }
-        this.isProcessing = false;
-    }
-}
 
 async function analyzeDownloadTask(list: IDictionary<PieceInfo> = selectList) {
     let size = list.size
